// src/lib/schemas/bgm-public-api.d.ts
export interface paths {
// ...
    "/v0/users/{username}/collections": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * 获取用户收藏
         * @description 获取对应用户的收藏，查看私有收藏需要access token。
         */
        get: operations["getUserCollectionsByUsername"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        UserSubjectCollection: {
            /**
             * Subject ID
             * @example 8
             */
            subject_id: number;
            subject_type: components["schemas"]["SubjectType"];
            /**
             * Rate
             * @example 4
             */
            rate: number;
            type: components["schemas"]["SubjectCollectionType"];
            /**
             * Comment
             * @example 看看
             */
            comment?: string;
            /**
             * Tags
             * @example [
             *       "柯南",
             *       "万年小学生",
             *       "推理",
             *       "青山刚昌",
             *       "TV"
             *     ]
             */
            tags: string[];
            /**
             * Ep Status
             * @example 5
             */
            ep_status: number;
            /**
             * Vol Status
             * @example 0
             */
            vol_status: number;
            /**
             * Updated At
             * Format: date-time
             * @description 本时间并不代表条目的收藏时间。修改评分，评价，章节观看状态等收藏信息时未更新此时间是一个 bug。请不要依赖此特性
             * @example 2022-06-19T18:44:13.6140127+08:00
             */
            updated_at: string;
            /** Private */
            private: boolean;
            subject?: components["schemas"]["SlimSubject"];
        };
// ...
    };
// ...
}